create constants file and refactor chessboard
