+++ src/components/Referee.ts
@@ -1,4 +1,4 @@
-import { PieceType, TeamType, Piece } from './Chessboard'
+import { PieceType, TeamType, Piece } from './Constants'
 
 export default class Referee {
     tileIsOccupied(x: number, y: number, boardState: Piece[]): boolean {
